# Optimize performance of ConvNeXt models