I've found the current ConvNeXt implementation is significantly slower than the official implementation. 

I've identified two patches that will achieve parity (in terms of speed):

1. LayerNorm2d: remove the `F.layer_norm` call and always use the custom implementation.
2. ConvNeXtBlock: `x = self.drop_path(x) + shortcut` -> `x = shortcut + self.drop_path(x)`.

Both patches are required to get the speed boost.

Comparison between the original and the patched version:

* [conv_tiny](https://colab.research.google.com/drive/1ZD6uCyQzYgF_8tLRBPr5rlVts1LFg7vt?usp=sharing) (2min 31s -> 1min 6s on T4)
* [conv_base](https://colab.research.google.com/drive/1z-xmrm-gYSbPCKOnLymOAy8470ns5aNG?usp=sharing) (2min 40s -> 1min 49s on P100)
* [conv_base](https://colab.research.google.com/drive/1eqdDkea3Xesls7cU_RfdBdAUBNAMVe2T?usp=sharing) (5min 13s -> 2min 53s on T4)

Ablation:
* [Without the LayerNorm patch](https://colab.research.google.com/drive/1WDWpYPWOGjz2xDy6mq5uS2H50TFcgqQh?usp=sharing)
* [Without the ConvNeXtBlock patch](https://colab.research.google.com/drive/1rv8azxg-kEY4aLYNWSBKyKvdh8owMWg0?usp=sharing)