@rwightman 2022-03-29
@ceshine your benchmark is on a P100, this is extremely GPU dependent and it's hard to find a good balance, my alterations were made because it was faster (for smaller models) on Ampere cards.