@ceshine 2022-03-30
@rwightman I see. FYI, My benchmark also includes T4 (Turing). My local testing on RTX 2070 yields the same results. 

I don't have access to Ampere cards, so sorry I wasn't aware that your alterations are faster on them.

I'm closing this PR, then.