@rwightman 2022-03-30
@ceshine overall the LayerNorm in convnext is a pretty large source of performance issues, especially when trying to use the optimal memory layout for conv nets (channels_last) in mixed precision or fp16. The variations across different hardware / CUDA/cuDNN / pytorch versions is also more significant than most models... :(